SX127x - sync with LoRaLib v4.0.0

+++ src/modules/SX1272.cpp
@@ -4,9 +4,9 @@ SX1272::SX1272(Module* mod) : SX127x(mod) {
   
 }
 
-uint8_t SX1272::begin(float freq, float bw, uint8_t sf, uint8_t cr, uint8_t syncWord, int8_t power) {
+int16_t SX1272::begin(float freq, float bw, uint8_t sf, uint8_t cr, uint8_t syncWord, int8_t power) {
   // execute common part
-  uint8_t state = SX127x::begin(SX1272_CHIP_VERSION, syncWord);
+  int16_t state = SX127x::begin(SX1272_CHIP_VERSION, syncWord);
   if(state != ERR_NONE) {
     return(state);
   }
@@ -46,7 +46,7 @@ uint8_t SX1272::begin(float freq, float bw, uint8_t sf, uint8_t cr, uint8_t sync
   return(ERR_NONE);
 }
 
-uint8_t SX1272::setFrequency(float freq) {
+int16_t SX1272::setFrequency(float freq) {
   // check frequency range
   if((freq < 860.0) || (freq > 1020.0)) {
     return(ERR_INVALID_FREQUENCY);
@@ -56,7 +56,7 @@ uint8_t SX1272::setFrequency(float freq) {
   return(SX127x::setFrequencyRaw(freq));
 }
 
-uint8_t SX1272::setBandwidth(float bw) {
+int16_t SX1272::setBandwidth(float bw) {
   uint8_t newBandwidth;
   
   // check alowed bandwidth values
@@ -71,7 +71,7 @@ uint8_t SX1272::setBandwidth(float bw) {
   }
   
   // set bandwidth and if successful, save the new setting
-  uint8_t state = SX1272::setBandwidthRaw(newBandwidth);
+  int16_t state = SX1272::setBandwidthRaw(newBandwidth);
   if(state == ERR_NONE) {
     SX127x::_bw = bw;
   }
@@ -79,7 +79,7 @@ uint8_t SX1272::setBandwidth(float bw) {
   return(state);
 }
 
-uint8_t SX1272::setSpreadingFactor(uint8_t sf) {
+int16_t SX1272::setSpreadingFactor(uint8_t sf) {
   uint8_t newSpreadingFactor;
   
   // check allowed spreading factor values
@@ -110,7 +110,7 @@ uint8_t SX1272::setSpreadingFactor(uint8_t sf) {
   }
   
   // set spreading factor and if successful, save the new setting
-  uint8_t state = SX1272::setSpreadingFactorRaw(newSpreadingFactor);
+  int16_t state = SX1272::setSpreadingFactorRaw(newSpreadingFactor);
   if(state == ERR_NONE) {
     SX127x::_sf = sf;
   }
@@ -118,7 +118,7 @@ uint8_t SX1272::setSpreadingFactor(uint8_t sf) {
   return(state);
 }
 
-uint8_t SX1272::setCodingRate(uint8_t cr) {
+int16_t SX1272::setCodingRate(uint8_t cr) {
   uint8_t newCodingRate;
   
   // check allowed coding rate values
@@ -140,7 +140,7 @@ uint8_t SX1272::setCodingRate(uint8_t cr) {
   }
   
   // set coding rate and if successful, save the new setting
-  uint8_t state = SX1272::setCodingRateRaw(newCodingRate);
+  int16_t state = SX1272::setCodingRateRaw(newCodingRate);
   if(state == ERR_NONE) {
     SX127x::_cr = cr;
   }
@@ -148,7 +148,7 @@ uint8_t SX1272::setCodingRate(uint8_t cr) {
   return(state);
 }
 
-uint8_t SX1272::setOutputPower(int8_t power) {
+int16_t SX1272::setOutputPower(int8_t power) {
   // check allowed power range
   if((power < -1) || (power > 20)) {
     return(ERR_INVALID_OUTPUT_POWER);
@@ -157,7 +157,7 @@ uint8_t SX1272::setOutputPower(int8_t power) {
   // set mode to standby
   SX127x::standby();
   
-  uint8_t state;
+  int16_t state;
   if(power < 15) {
     // power is less than 15 dBm, enable PA0 on RFIO
     state = _mod->SPIsetRegValue(SX127X_REG_PA_CONFIG, SX127X_PA_SELECT_RFO, 7, 7);
@@ -178,7 +178,7 @@ uint8_t SX1272::setOutputPower(int8_t power) {
   return(state);
 }
 
-uint8_t SX1272::setBandwidthRaw(uint8_t newBandwidth) {
+int16_t SX1272::setBandwidthRaw(uint8_t newBandwidth) {
   // set mode to standby
   SX127x::standby();
   
@@ -186,12 +186,12 @@ uint8_t SX1272::setBandwidthRaw(uint8_t newBandwidth) {
   return(_mod->SPIsetRegValue(SX127X_REG_MODEM_CONFIG_1, newBandwidth, 7, 6));
 }
 
-uint8_t SX1272::setSpreadingFactorRaw(uint8_t newSpreadingFactor) {
+int16_t SX1272::setSpreadingFactorRaw(uint8_t newSpreadingFactor) {
   // set mode to standby
   SX127x::standby();
   
   // write registers
-  uint8_t state = 0;
+  int16_t state = 0;
   if(newSpreadingFactor == SX127X_SF_6) {
     state |= _mod->SPIsetRegValue(SX127X_REG_MODEM_CONFIG_1, SX1272_HEADER_IMPL_MODE | SX1272_RX_CRC_MODE_OFF, 2, 1);
     state |= _mod->SPIsetRegValue(SX127X_REG_MODEM_CONFIG_2, SX127X_SF_6 | SX127X_TX_MODE_SINGLE, 7, 3);
@@ -207,7 +207,7 @@ uint8_t SX1272::setSpreadingFactorRaw(uint8_t newSpreadingFactor) {
   return(state);
 }
 
-uint8_t SX1272::setCodingRateRaw(uint8_t newCodingRate) {
+int16_t SX1272::setCodingRateRaw(uint8_t newCodingRate) {
   // set mode to standby
   SX127x::standby();
   
@@ -215,9 +215,9 @@ uint8_t SX1272::setCodingRateRaw(uint8_t newCodingRate) {
   return(_mod->SPIsetRegValue(SX127X_REG_MODEM_CONFIG_1, newCodingRate, 5, 3));
 }
 
-uint8_t SX1272::config() {
+int16_t SX1272::config() {
   // configure common registers
-  uint8_t state = SX127x::config();
+  int16_t state = SX127x::config();
   if(state != ERR_NONE) {
     return(state);
   }
