+++ src/modules/SX1276.cpp
@@ -4,7 +4,7 @@ SX1276::SX1276(Module* mod) : SX1278(mod) {
   
 }
 
-uint8_t SX1276::setFrequency(float freq) {
+int16_t SX1276::setFrequency(float freq) {
   // check frequency range
   if((freq < 137.0) || (freq > 1020.0)) {
     return(ERR_INVALID_FREQUENCY);
