+++ src/modules/SX1277.h
@@ -10,8 +10,8 @@ class SX1277: public SX1278 {
     SX1277(Module* mod);
     
     // configuration methods
-    uint8_t setFrequency(float freq);
-    uint8_t setSpreadingFactor(uint8_t sf);
+    int16_t setFrequency(float freq);
+    int16_t setSpreadingFactor(uint8_t sf);
 };
 
 #endif
