+++ src/modules/SX1279.cpp
@@ -4,7 +4,7 @@ SX1279::SX1279(Module* mod) : SX1278(mod) {
   
 }
 
-uint8_t SX1279::setFrequency(float freq) {
+int16_t SX1279::setFrequency(float freq) {
   // check frequency range
   if((freq < 137.0) || (freq > 960.0)) {
     return(ERR_INVALID_FREQUENCY);
